Move MapTypeIdValuesArray from map-type-ids.ts to diagram-type-ids.ts, and delete map-type-ids.ts
