--- projects/fhi-angular-highcharts/src/lib/constants-and-enums/map-type-ids.ts
@@ -1,3 +1,0 @@
-import { MapTypeIdValues } from './diagram-type-ids';
-
-export const MapTypeIdValuesArray = [MapTypeIdValues.mapFylker, MapTypeIdValues.mapFylker2019];
